requirements.txt
django
djangorestframework
django-cors-headers
dj-database-url
djangorestframework>=3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn
psycopg2-binary
pillow
django-extensions>=4.1
pytest
pytest-django
requests>=2.32.3
stripe
python-pptx
pdf2image
reportlab>=4.0.0
weasyprint>=60.0
numpy>=2.0.0
matplotlib>=3.7.0
social-auth-app-django
python-jose[cryptography]
PyJWT[crypto]
cryptography
python-dotenv
auth0-python
bleach>=6.2.0

# Email Integration Libraries
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
msal>=1.24.0
exchangelib>=5.0.0
imapclient>=3.0.0
secure-smtplib
email-validator

# AI Integration Libraries
openai>=1.3.0

# Background Task Processing Libraries
celery>=5.3.0
redis>=4.5.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
flower>=2.0.0

# AWS S3 Integration Libraries
boto3>=1.34.0
botocore>=1.34.0

# Phase 4: Performance & Optimization dependencies
django-redis>=5.3.0
elasticsearch>=8.9.0
psutil>=5.9.0
django-compressor>=4.4
django-filter
pdfcrowd>=5.17.0
